# swagger.yaml
openapi: 3.1.0
info:
  title: API Documentation AND103
  version: 1.0.0
  description: Tài liệu API quản lý Sản phẩm, Danh mục và Hình ảnh
servers:
  - url: http://localhost:3001
    description: Local Development Server

# --- KHU VỰC ĐỊNH NGHĨA MODEL CHUNG (Schema) ---
components:
  schemas:
...
    ProductInput:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: "Macbook Air M2"
        description:
          type: string
          example: "Máy tính xách tay Apple..."
        price:
          type: number
          example: 25000000
        quantity:
          type: integer
          example: 100
        status:
          type: boolean
          example: true
        cateID:
          type: string
          description: ID của danh mục
          example: "654a..."

# --- KHU VỰC ĐỊNH NGHĨA API (Paths) ---
paths:
...
  /products/add:
    post:
      summary: Thêm sản phẩm mới
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        200:
          description: Thêm thành công
  /products/update:
    put:
      summary: Cập nhật sản phẩm
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  properties:
                    id:
                      type: string
                      description: ID sản phẩm (Bắt buộc)
                - $ref: '#/components/schemas/ProductInput'
      responses:
        200:
          description: Cập nhật thành công
  /products/delete/{id}:
    delete:
      summary: Xóa sản phẩm theo ID
      tags: [Products]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID sản phẩm cần xóa
      responses:
        200:
          description: Xóa thành công

  # --- CÁC API GET (CÂU HỎI MẪU) ---
  # Đã gộp tag thành [Products]
...
  /products/cau15/{cateId}:
    get:
      summary: Câu 15 - Lấy dssp theo cateID (loại bỏ status false)
      tags: [Products]
      parameters:
        - in: path
          name: cateId
          schema:
            type: string
          required: true
          description: ID danh mục muốn lọc
      responses:
        200:
          description: Thành công
  /products/cau20/{cateId}:
    get:
      summary: Câu 20 - Lấy dssp theo cateID (chỉ lấy name, price, quantity)
      tags: [Products]
      parameters:
        - in: path
          name: cateId
          schema:
            type: string
          required: true
      responses:
        200:
          description: Thành công

  # ================= USER API =================
...